requirements.txt
streamlit
pillow
numpy
tensorflow
scikit-learn
matplotlib